+++ src/ownedchain.rs
@@ -151,7 +151,7 @@ impl OwnedChain {
     }
 
     /// Number of blocks in this owned chain.
-    pub fn count(&self) -> u128 {
+    pub fn count(&self) -> u64 {
         self.chain.count()
     }
